+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/delivery/controller/DeliveryRequestApiController.java
@@ -1,5 +1,7 @@
 /*
  * Copyright (c) 2024 Volkswagen AG
+ * Copyright (c) 2025 Fraunhofer-Gesellschaft zur Foerderung der angewandten Forschung e.V.
+ * (represented by Fraunhofer ISST)
  * Copyright (c) 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
@@ -29,26 +31,32 @@ import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
 import org.eclipse.tractusx.puris.backend.delivery.logic.dto.deliverysamm.DeliveryInformation;
 import org.eclipse.tractusx.puris.backend.delivery.logic.service.DeliveryRequestApiService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.regex.Pattern;
 
-@RestController
-@RequestMapping("delivery-information")
-@Slf4j
 /**
  * This class offers the endpoint for requesting the PlannedProduction Submodel 2.0.0
  */
+@RestController
+@RequestMapping("delivery-information")
+@Slf4j
 public class DeliveryRequestApiController {
 
     @Autowired
-    private DeliveryRequestApiService deliveryRequestApiSrvice;
+    private DeliveryRequestApiService deliveryRequestApiService;
 
     private final Pattern bpnlPattern = PatternStore.BPNL_PATTERN;
 
     private final Pattern urnPattern = PatternStore.URN_OR_UUID_PATTERN;
 
+    @RequestMapping(value = "/**")
+    @ResponseStatus(HttpStatus.NOT_IMPLEMENTED)
+    public ResponseEntity<String> handleNotImplemented() {
+        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
+    }
 
     @Operation(summary = "This endpoint receives the Delivery Information Submodel 2.0.0 requests. " +
         "This endpoint is meant to be accessed by partners via EDC only. ")
@@ -58,7 +66,7 @@ public class DeliveryRequestApiController {
         @ApiResponse(responseCode = "500", description = "Internal Server Error", content = @Content),
         @ApiResponse(responseCode = "501", description = "Unsupported representation", content = @Content)
     })
-    @GetMapping("request/{materialNumberCx}/{representation}")
+    @GetMapping("request/{materialNumberCx}/submodel/{representation}")
     public ResponseEntity<DeliveryInformation> getDeliveryMapping(
         @RequestHeader("edc-bpn") String bpnl,
         @PathVariable String materialNumberCx,
@@ -79,7 +87,7 @@ public class DeliveryRequestApiController {
         }
 
         log.info("Received request for " + materialNumberCx + " from " + bpnl);
-        var samm = deliveryRequestApiSrvice.handleDeliverySubmodelRequest(bpnl, materialNumberCx);
+        var samm = deliveryRequestApiService.handleDeliverySubmodelRequest(bpnl, materialNumberCx);
         if (samm == null) {
             log.error("SAMM for delivery is null, return 500.");
             return ResponseEntity.status(500).build();