+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/masterdata/controller/MaterialController.java
@@ -31,6 +31,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
 import org.eclipse.tractusx.puris.backend.masterdata.domain.model.Material;
 import org.eclipse.tractusx.puris.backend.masterdata.logic.dto.MaterialEntityDto;
+import org.eclipse.tractusx.puris.backend.masterdata.logic.service.MaterialRefreshService;
 import org.eclipse.tractusx.puris.backend.masterdata.logic.service.MaterialService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -50,6 +51,9 @@ public class MaterialController {
 
     @Autowired
     private MaterialService materialService;
+
+    @Autowired
+    private MaterialRefreshService materialRefreshService;
 
     @Autowired
     private Validator validator;
@@ -157,5 +161,16 @@ public class MaterialController {
         return new ResponseEntity<>(materialService.findAll().
             stream().map(x -> modelMapper.map(x, MaterialEntityDto.class)).collect(Collectors.toList()),
             HttpStatusCode.valueOf(200));
+    }
+
+    @GetMapping("/refresh")
+    @Operation(description = "Requests material data for the given material from all partners")
+    public ResponseEntity<String> refreshMaterialData(@RequestParam @Parameter(description = "encoded in base64") String ownMaterialNumber) {
+        ownMaterialNumber = new String(Base64.getDecoder().decode(ownMaterialNumber));
+        if (!materialPattern.matcher(ownMaterialNumber).matches()) {
+            return new ResponseEntity<>(HttpStatusCode.valueOf(400));
+        }
+        materialRefreshService.refreshPartnerData(ownMaterialNumber);
+        return new ResponseEntity<>(ownMaterialNumber, HttpStatusCode.valueOf(200));
     }
 }