+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/production/controller/ProductionRequestApiController.java
@@ -1,5 +1,7 @@
 /*
  * Copyright (c) 2024 Volkswagen AG
+ * Copyright (c) 2025 Fraunhofer-Gesellschaft zur Foerderung der angewandten Forschung e.V.
+ * (represented by Fraunhofer ISST)
  * Copyright (c) 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
@@ -29,17 +31,18 @@ import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
 import org.eclipse.tractusx.puris.backend.production.logic.dto.plannedproductionsamm.PlannedProductionOutput;
 import org.eclipse.tractusx.puris.backend.production.logic.service.ProductionRequestApiService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.regex.Pattern;
 
-@RestController
-@RequestMapping("planned-production")
-@Slf4j
 /**
  * This class offers the endpoint for requesting the PlannedProduction Submodel 2.0.0
  */
+@RestController
+@RequestMapping("planned-production")
+@Slf4j
 public class ProductionRequestApiController {
 
     @Autowired
@@ -49,6 +52,11 @@ public class ProductionRequestApiController {
 
     private final Pattern urnPattern = PatternStore.URN_OR_UUID_PATTERN;
 
+    @RequestMapping(value = "/**")
+    @ResponseStatus(HttpStatus.NOT_IMPLEMENTED)
+    public ResponseEntity<String> handleNotImplemented() {
+        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
+    }
 
     @Operation(summary = "This endpoint receives the PlannedProduction Submodel 2.0.0 requests. " +
         "This endpoint is meant to be accessed by partners via EDC only. ")
@@ -58,7 +66,7 @@ public class ProductionRequestApiController {
         @ApiResponse(responseCode = "500", description = "Internal Server Error", content = @Content),
         @ApiResponse(responseCode = "501", description = "Unsupported representation", content = @Content)
     })
-    @GetMapping("request/{materialnumbercx}/{representation}")
+    @GetMapping("request/{materialnumbercx}/submodel/{representation}")
     public ResponseEntity<PlannedProductionOutput> getProductionMapping(
         @RequestHeader("edc-bpn") String bpnl,
         @PathVariable String materialnumbercx,
@@ -70,9 +78,6 @@ public class ProductionRequestApiController {
         }
         if (!"$value".equals(representation)) {
             log.warn("Rejecting request at PlannedProduction Submodel request 2.0.0 endpoint, missing '$value' in request");
-            if (!PatternStore.NON_EMPTY_NON_VERTICAL_WHITESPACE_PATTERN.matcher(representation).matches()) {
-                representation = "<REPLACED_INVALID_REPRESENTATION>";
-            }
             return ResponseEntity.status(501).build();
         }
         var samm = productionRequestApiService.handleProductionSubmodelRequest(bpnl, materialnumbercx);